feature: amazing Icons namespace pattern created 😍
