+++ docker-compose.yml
@@ -8,6 +8,7 @@ services:
       - '3000:3000'
       - '5000:5000'
     volumes:
+      - './package.json:/usr/src/app/node_modules/svg-symbol-sprite-loader/package.json'
       - './src:/usr/src/app/node_modules/svg-symbol-sprite-loader/src'
       - './test-app/src:/usr/src/app/src'
       - './test-app/webpack.config.js:/usr/src/app/webpack.config.js'
